[REFACTOR] #155 - 알림 기능 디렉토리 생성 및 코드 이동
